Isolando o interpretador de funções
