+++ rinha-compiler/CMakeLists.txt
@@ -24,6 +24,7 @@ add_library(${PROJECT_NAME}
                     "${CMAKE_CURRENT_SOURCE_DIR}/include/parser_json_to_ast.hpp"
                     "${CMAKE_CURRENT_SOURCE_DIR}/include/walker.hpp"
                     "${CMAKE_CURRENT_SOURCE_DIR}/include/interpreter.hpp"
+                    "${CMAKE_CURRENT_SOURCE_DIR}/include/interpreter_function.hpp"
                     "${CMAKE_CURRENT_SOURCE_DIR}/include/interpreter_binary_operation.hpp"
                     "${CMAKE_CURRENT_SOURCE_DIR}/include/compiler.hpp"
                     "${CMAKE_CURRENT_SOURCE_DIR}/src/stack.cpp"
@@ -33,6 +34,7 @@ add_library(${PROJECT_NAME}
                     "${CMAKE_CURRENT_SOURCE_DIR}/src/symbol_table.cpp"
                     "${CMAKE_CURRENT_SOURCE_DIR}/src/parser_json_to_ast.cpp"
                     "${CMAKE_CURRENT_SOURCE_DIR}/src/interpreter_binary_operation.cpp"
+                    "${CMAKE_CURRENT_SOURCE_DIR}/src/interpreter_function.cpp"
                     "${CMAKE_CURRENT_SOURCE_DIR}/src/interpreter_walker.cpp"
                     "${CMAKE_CURRENT_SOURCE_DIR}/src/interpreter.cpp"
                     "${CMAKE_CURRENT_SOURCE_DIR}/src/compiler.cpp"
