+++ rinha-compiler/src/interpreter_walker.cpp
@@ -3,6 +3,7 @@
 #include <rinha_exception.hpp>
 #include <type.hpp>
 #include <interpreter_binary_operation.hpp>
+#include <interpreter_function.hpp>
 
 rinha_compiler::stack _stack;
 rinha_compiler::Memory _memory;
@@ -18,145 +19,6 @@ using namespace rinha_compiler::interpreter;
     }
     void run_str(Str& term){
         _stack.push(term.value);
-    }
-
-    void load_paran_list(box<Function>& function){
-        auto parameters = function->parameters;
-        for (int i = parameters.size() - 1; i >= 0; i--)
-        {
-            Type value = _stack.pop();
-            std::string id = parameters[i].text;
-            rinha_compiler::Symbol symbol = rinha_compiler::init_symbol(id, current_scope);
-
-            symbol = symbolTable->Put(id, symbol);
-
-            _memory.store(symbol, value);
-        }
-    }
-
-    void run_call(box<Call>& term){
-        int saved_scope = current_scope;
-
-        //Carrega argumentos na pilha
-        int arity = term->arguments.terms.size();
-        if(arity > 0){
-            for(auto arg: term->arguments.terms)
-                std::visit(walker::VisitTerm{}, arg);
-        }
-
-        //Recupera nome da função
-        std::visit(walker::VisitTerm{}, term->callee.terms.front());
-        Type function_name = _stack.pop();
-        if(!std::holds_alternative<std::string>(function_name))
-            throw rinha_compiler::RinhaException("Nome de função inválido.", term->location);
-        
-        //Recupera função da tabela de simbolos
-        std::string id = std::get<std::string>(function_name);
-        rinha_compiler::Symbol symbol = symbolTable->Get(id);
-        
-        //Muda de escopo
-        //A função sempre é salva com o espopo da parent. Então o novo escopo é:
-        current_scope = symbol.scope + 1;
-
-        //Aloca mas espaço na memória, para o novo escopo
-        _memory.increment(current_scope + 1);
-        //Cria página de memória
-        _memory.push(current_scope);
-        
-        //Busca e aponta para o bloco da função na AST
-        auto let_function = std::get<box<Let>>(symbol.term);
-
-        //Varre o bloco da função recursivamente
-        std::visit(walker::VisitTerm{}, let_function->value.terms.front());
-
-        //Deleta página de memória
-        _memory.pop(current_scope);
-
-        current_scope = saved_scope;
-    }
-    void run_binary(box<Binary>& term){
-        
-        switch (term->op)
-        {
-            case BinaryOp::Add:
-                trace("run_plus");
-                run_plus(term);
-                break;
-            case BinaryOp::Sub:
-                trace("run_minus");
-                run_minus(term);
-                break;
-            case BinaryOp::Mul:
-                trace("run_times");
-                run_times(term);
-                break;
-            case BinaryOp::Div:
-                trace("run_over");
-                run_over(term);
-                break;
-            case BinaryOp::Rem:
-                trace("run_rem");
-                run_rem(term);
-                break;
-            case BinaryOp::Eq:
-                trace("run_eq");
-                run_eq(term);
-                break;
-            case BinaryOp::Neq:
-                trace("run_neq");
-                run_neq(term);
-                break;
-            case BinaryOp::Lt:
-                trace("run_lt");
-                run_lt(term);
-                break;
-            case BinaryOp::Gt:
-                trace("run_gt");
-                run_gt(term);
-                break;
-            case BinaryOp::Lte:
-                trace("run_lte");
-                run_lte(term);
-                break;
-            case BinaryOp::Gte:
-                trace("run_gte");
-                run_gte(term);
-                break;
-            case BinaryOp::And:
-                trace("run_and");
-                run_and(term);
-                break;
-            case BinaryOp::Or:
-                trace("run_or");
-                run_or(term);
-                break;
-            default:
-                throw rinha_compiler::RinhaException("Operação binária inválida", term->location);
-                break;
-        }
-    }
-    void run_function(box<Function>& term){
-        rinha_compiler::SymbolTable* symbol_table_function = new rinha_compiler::SymbolTable(symbolTable);
-        symbolTable = symbol_table_function;
-        load_paran_list(term);
-
-        std::visit(walker::VisitTerm{}, term->value.terms.front());
-
-        symbolTable = symbol_table_function->GetParent();
-    }
-
-    void run_function_decl(box<Let>& term){
-        std::string id = term->name.text;
-        Term value = term->value.terms.front();
-        //current_scope++;
-        _memory.increment(current_scope + 1);
-
-        rinha_compiler::Symbol symbol = rinha_compiler::init_symbol(id, current_scope, term);
-
-        symbol = symbolTable->Put(id, symbol);
-        _memory.store(symbol, id);
-
-        //current_scope--;
     }
 
      void run_var_decl(box<Let>& term){
